refactor: modularize admin scripts
